# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API manages Quality of Service (QoS) rules using Linux Traffic Control (tc).",
        "title": "QoS Traffic Control API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/qos/htb/global/limit": {
            "post": {
                "description": "Changes the rate/ceil of the HTB root class (1:1) on both LAN and WAN interfaces.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HTB Setup"
                ],
                "summary": "Updates the global rate limit for the existing HTB structure.",
                "parameters": [
                    {
                        "description": "Interface and Bandwidth Update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.RuleUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status: HTB global limit updated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error: Invalid request format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error: Failed to update HTB limit",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/qos/ip/limit": {
            "post": {
                "description": "Adds a child class under the global HTB qdisc and applies a filter to match the specified IP address.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HTB IP Control"
                ],
                "summary": "Applies a specific HTB rate limit to a client IP.",
                "parameters": [
                    {
                        "description": "IP and Rate Limit Configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.IPControlRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status: IP rate limit applied successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error: Invalid request format or missing field",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error: Failed to apply IP rate limit",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/qos/ip/remove": {
            "post": {
                "description": "Removes the specific child class and filter associated with the IP.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HTB IP Control"
                ],
                "summary": "Removes the HTB rate limit applied to a client IP.",
                "parameters": [
                    {
                        "description": "IP address to remove limit from",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.IPControlRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status: IP rate limit removed successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error: Invalid request format or missing field",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error: Failed to remove IP rate limit",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/qos/reset": {
            "post": {
                "description": "Deletes the root qdisc on both the specified LAN and WAN interfaces, effectively removing all HTB or TBF rules.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cleanup"
                ],
                "summary": "Removes all shaping (QDiscs) from both interfaces.",
                "parameters": [
                    {
                        "description": "Interfaces to reset",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SetupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status: Shaping successfully reset on both interfaces",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error: Invalid request format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error: Failed to reset shaping",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/qos/setup": {
            "post": {
                "description": "Sets up the root HTB qdisc and class 1:1 on both LAN and WAN interfaces.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HTB Setup"
                ],
                "summary": "Initializes the HTB (Hierarchical Token Bucket) structure.",
                "parameters": [
                    {
                        "description": "Interface and Global Bandwidth Configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SetupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status: HTB structure successfully initialized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error: Invalid request format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error: Failed to set up HTB structure",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/qos/simple/limit": {
            "post": {
                "description": "Applies a simple Token Bucket Filter (TBF) qdisc to the root of both LAN and WAN interfaces.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Simple TBF"
                ],
                "summary": "Applies a simple interface-wide limit using TBF.",
                "parameters": [
                    {
                        "description": "Interface and Bandwidth/Latency parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.RuleUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status: Simple TBF rule applied successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error: Invalid request format or missing field",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error: Failed to apply TBF rule",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/qos/stats": {
            "get": {
                "description": "Lit les compteurs de /proc/net/dev et calcule le débit (Rx et Tx) en Mbps en utilisant l'état précédent stocké côté serveur.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitoring"
                ],
                "summary": "Récupère le débit de trafic par interface en une seule fois (polling HTTP).",
                "responses": {
                    "200": {
                        "description": "Débit de trafic par interface",
                        "schema": {
                            "$ref": "#/definitions/handler.InterfaceStats"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la récupération des statistiques",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/qos/stream": {
            "get": {
                "description": "Met à niveau la connexion HTTP vers WebSocket et pousse les statistiques de débit par IP et globales (issues du QoSManager) toutes les ~2 secondes.",
                "tags": [
                    "Monitoring"
                ],
                "summary": "Ouvre une connexion WebSocket pour streamer le débit de trafic par IP en temps réel.",
                "responses": {}
            }
        },
        "/status/lanips": {
            "get": {
                "description": "Uses 'ip neighbor show' to find IPs in REACHABLE, STALE, or DELAY state on the local network interface.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Status"
                ],
                "summary": "Retrieves connected client IPs on the LAN interface.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The LAN interface name (e.g., eth0)",
                        "name": "lan_interface",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of connected IPv4 or IPv6 (non-fe80) addresses",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error: Missing or invalid lan_interface parameter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error: Failed to execute ip neighbor command",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.IPControlRequest": {
            "type": "object",
            "required": [
                "ip",
                "lan_interface",
                "wan_interface"
            ],
            "properties": {
                "ip": {
                    "type": "string",
                    "example": "192.168.1.10"
                },
                "lan_interface": {
                    "type": "string",
                    "example": "eth0"
                },
                "rate_limit": {
                    "type": "string",
                    "example": "5mbit"
                },
                "wan_interface": {
                    "type": "string",
                    "example": "wlan1"
                }
            }
        },
        "handler.IPTrafficStat": {
            "type": "object",
            "properties": {
                "download_mbps": {
                    "type": "number"
                },
                "ip": {
                    "type": "string"
                },
                "is_limited": {
                    "type": "boolean"
                },
                "status": {
                    "description": "e.g., \"Active\", \"New\", \"Disconnected\"",
                    "type": "string"
                },
                "upload_mbps": {
                    "type": "number"
                }
            }
        },
        "handler.InterfaceStats": {
            "type": "object",
            "properties": {
                "current_rate_rx_mbps": {
                    "type": "number",
                    "example": 5.5
                },
                "current_rate_tx_mbps": {
                    "type": "number",
                    "example": 12.8
                },
                "interface_name": {
                    "type": "string",
                    "example": "eth0"
                },
                "ip_stats": {
                    "description": "Empty for now",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/handler.IPTrafficStat"
                    }
                },
                "total_bytes_rx": {
                    "type": "integer",
                    "example": 1100000
                },
                "total_bytes_tx": {
                    "type": "integer",
                    "example": 2351000
                }
            }
        },
        "handler.RuleUpdateRequest": {
            "type": "object",
            "required": [
                "lan_interface",
                "rate_limit",
                "wan_interface"
            ],
            "properties": {
                "lan_interface": {
                    "type": "string",
                    "example": "eth0"
                },
                "latency": {
                    "type": "string",
                    "example": "50ms"
                },
                "rate_limit": {
                    "type": "string",
                    "example": "10mbit"
                },
                "wan_interface": {
                    "type": "string",
                    "example": "wlan1"
                }
            }
        },
        "handler.SetupRequest": {
            "type": "object",
            "required": [
                "lan_interface",
                "total_bandwidth",
                "wan_interface"
            ],
            "properties": {
                "lan_interface": {
                    "type": "string",
                    "example": "eth0"
                },
                "total_bandwidth": {
                    "type": "string",
                    "example": "10mbit"
                },
                "wan_interface": {
                    "type": "string",
                    "example": "wlan1"
                }
            }
        }
    }
}